+++ src/libcsg/modules/io/lammpsdumpwriter.cc
@@ -35,7 +35,7 @@ void LAMMPSDumpWriter::Close() { fclose(_out); }
 void LAMMPSDumpWriter::Write(Topology *conf) {
   Topology *top = conf;
   Eigen::Matrix3d box = conf->getBox();
-  fprintf(_out, "ITEM: TIMESTEP\n%i\n", top->getStep());
+  fprintf(_out, "ITEM: TIMESTEP\n%ld\n", top->getStep());
   fprintf(_out, "ITEM: NUMBER OF ATOMS\n%i\n", (int)top->Beads().size());
   fprintf(_out, "ITEM: BOX BOUNDS pp pp pp\n");
   fprintf(_out, "0 %f\n0 %f\n0 %f\n", box(0, 0) * conv::nm2ang,
@@ -56,9 +56,9 @@ void LAMMPSDumpWriter::Write(Topology *conf) {
        iter != conf->Beads().end(); ++iter) {
     Bead *bi = *iter;
 
-    int type_id = conf->getBeadTypeId(bi->getType());
+    long int type_id = conf->getBeadTypeId(bi->getType());
 
-    fprintf(_out, "%i %i", bi->getId() + 1, type_id);
+    fprintf(_out, "%ld %li", bi->getId() + 1, type_id);
     fprintf(_out, " %f %f %f", bi->getPos().x() * conv::nm2ang,
             bi->getPos().y() * conv::nm2ang, bi->getPos().z() * conv::nm2ang);
     if (v) {
